rename newtimer to timer since we no longer have the old test
