+++ dev/com.ibm.ws.cdi.ejb_fat/fat/src/com/ibm/ws/cdi/ejb/apps/timer/EjbSessionBean.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.cdi.ejb.apps.newtimer;
+package com.ibm.ws.cdi.ejb.apps.timer;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
