+++ dev/com.ibm.ws.cdi.ejb_fat/fat/src/com/ibm/ws/cdi/ejb/apps/timer/EjbSessionBeanLocal.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015 IBM Corporation and others.
+ * Copyright (c) 2022 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,22 +8,9 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.cdi.ejb.apps.newtimer;
+package com.ibm.ws.cdi.ejb.apps.timer;
 
-import static java.lang.annotation.ElementType.TYPE;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
+public interface EjbSessionBeanLocal {
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import javax.interceptor.InterceptorBinding;
-
-/**
- * CDI interceptor binding
- */
-@InterceptorBinding
-@Target({ TYPE })
-@Retention(RUNTIME)
-public @interface MyCDIInterceptorBinding {
-
+    void initTimer();
 }
