+++ dev/com.ibm.ws.cdi.ejb_fat/fat/src/com/ibm/ws/cdi/ejb/tests/FATSuite.java
@@ -31,7 +31,6 @@ import com.ibm.ws.fat.util.FatLogHandler;
                 EjbTimerTest.class,
                 InjectParameterTest.class,
                 MultipleNamedEJBTest.class,
-                NewEjbTimerTest.class,
                 StatefulSessionBeanInjectionTest.class,
 })
 public class FATSuite {
