+++ dev/com.ibm.ws.cdi.ejb_fat/test-applications/EjbTimer.war/resources/META-INF/permissions.xml
